requirements.txt
addml==0.0
aiofiles==0.4.0
aiohttp==3.6.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
audiomd==0.0
chardet==3.0.4
Click==7.0
coverage==4.5.4
dpres-signature==0.0
ffmpeg-python==0.2.0
file-magic==0.4.0
file-scraper==0.0
future==0.17.1
idna==2.8
importlib-metadata==0.23
lxml==4.4.1
M2Crypto==0.35.2
mets==0.0
more-itertools==7.2.0
multidict==4.5.2
nisomix==0.0
olefile==0.46
opf-fido==1.3.12
packaging==19.2
Pillow==6.0.0
pluggy==0.13.0
premis==0.0
py==1.8.0
pymediainfo==4.1
pyparsing==2.4.2
pytest==5.1.3
pytest-cov==2.7.1
python-dateutil==2.8.0
python-mimeparse==1.6.0
scandir==1.10.0
siptools==0.0
six==1.12.0
untangle==1.1.1
videomd==0.0
Wand==0.5.7
wcwidth==0.1.7
xml-helpers==0.0
yarl==1.3.0
zipp==0.6.0
